Revert edje calculate optimizations.
Edje is tricky, it's event processing is too weird and Cedric's
changes to make it work are not working as expected. Edje freezes
itself while processing signals, but in mouse down cb it forces
recalculate, which seems was previously ignored, but now they are not.

We should look at how to fix this and then re-apply this patch.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@35908 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33