* ... and finally save the generated EDC back to the edje file
This means that file created with the editor can now be decompiled!!

But make attention because not every feature of edje is supported.
This means that (if you modify an existing file) you will lost some feature
while saving; like MACRO, inherit values, comments in the EDC code and some 
other smallies.
But if you have created the file with the editor all will be edje_decc well.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@36537 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33