Edje_Edit API break.
Change group_del() to receive the name of the group to be deleted, and
change the function to not delete a group currently loaded. This causes
problems at the time of deleting the Evas_Object.
Also changed a bit the save() function and added save_all(), which saves
every group loaded, not only the one set to the object. This is mainly so
at the time of deleting a group, we can save the whole file and thus avoid
it getting out of sync with references if a group is deleted and the file
not saved afterwards.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@45720 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33