Edje entry: Remove the useless calls to "update_from_curs".
We don't really need or use the coordinates anywhere, we just saved them
and worked hard on maintaining them, but didn't really use them.
Now we only calculate them in the one specific case where we actually
need them. This speeds things up and also means the layouting may not be
called in some cases when there's no need to layout.

git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@59520 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33