From: Jihoon Kim <jihoon48.kim@samsung.com>
When elm_entry_entry_set() is called, the preedit text in the entry should
be cleared. ATM, it doesn’t works so.

This patch will resolve this problem.
 
When ecore_imf_context_reset() is called, the preedit text in the immodule
will be committed by calling ‘ecore_imf_context_commit_event_add()’.

ecore_imf_context_commit_event_add puts the event in ecore event queue, so
it will be processed in the idle state of ecore loop.

According to this mechanism, commit text is inserted later, so I use
‘ecore_main_loop_iterate();’ for inserting the commit string instantly.



git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@59732 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33