change text's single recalc to use apply version.
Original _edje_part_recalc_single_text() was not working as
expected since it was not doing size fit, range, ellipsis and so
on.

The purpose of this function compared with
_edje_text_recalc_apply() is to be faster, not calling Evas update
functions. However for text this is quite difficult given that to
fit we need to set the font, size, style, etc. If it was done
correctly, we'd save some calls to move and some color sets,
however those shouldn't matter much in the overall picture.

I've changed this to force applying the value, it should be more
correct and not so slow. The previous code is kept below for
reference but should be removed before next release!




git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@62640 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33