edje: add broadcast_signal: bool.
This make it possible to completly disable signal broadcasting as this
new behaviour broke Edje 1.0 file. It's also now possible to use the
same group in different part in the same parent group without any issue.

I am tempted to backport this patch to 1.1 branch as it would make it
play nicely with file coming from Edje 1.0.

Another issue that this patch fix is that I did increment the minor version
as we really have add a lot of addition since Edje 1.1 and Edje file build
with trunk may not play well anymore on Edje 1.1.


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@67936 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33