edje: preserve user box and table items during call of edje_object_file_set.
NOTE: now you can change theme dynamically in elementary apps more reliably.
This doesn't handle the case where the swallow was done in a parent object and
the reswallow should happen in a another group. I don't how to fix that use
case.
don't see yet how to handle that


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/edje@70901 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33